+++ src/DotPulsar/Internal/Consumer.cs
@@ -16,35 +16,27 @@ namespace DotPulsar.Internal;
 
 using DotPulsar.Abstractions;
 using DotPulsar.Exceptions;
-using DotPulsar.Extensions;
 using DotPulsar.Internal.Abstractions;
-using DotPulsar.Internal.Compression;
+using DotPulsar.Internal.Events;
 using DotPulsar.Internal.Extensions;
 using DotPulsar.Internal.PulsarApi;
+using Microsoft.Extensions.ObjectPool;
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Nito.AsyncEx;
 
-public sealed class Consumer<TMessage> : IConsumer<TMessage>, IRegisterEvent
+public sealed class Consumer<TMessage> : IContainsChannel, IConsumer<TMessage>
 {
+    private readonly Guid _correlationId;
+    private readonly IRegisterEvent _eventRegister;
+    private IConsumerChannel<TMessage> _channel;
+    private readonly ObjectPool<CommandAck> _commandAckPool;
     private readonly IExecute _executor;
-    private readonly StateManager<ConsumerState> _state;
-    private readonly ConsumerOptions<TMessage> _options;
-    private readonly ProcessManager _processManager;
-    private readonly IHandleException _exceptionHandler;
-    private readonly IConnectionPool _connectionPool;
-    private readonly CancellationTokenSource _cts;
-    private readonly ConcurrentDictionary<string, IConsumer<TMessage>> _consumers;
-    private readonly AsyncReaderWriterLock _lock;
-    private ConcurrentQueue<IMessage<TMessage>> _messagesQueue;
+    private readonly IStateChanged<ConsumerState> _state;
+    private readonly IConsumerChannelFactory<TMessage> _factory;
     private int _isDisposed;
-    private int _consumerCount;
-    private uint _numberOfPartitions;
-    private bool _isPartitioned;
     private Exception? _faultException;
 
     public Uri ServiceUrl { get; }
@@ -52,31 +44,29 @@ public sealed class Consumer<TMessage> : IConsumer<TMessage>, IRegisterEvent
     public string Topic { get; }
 
     public Consumer(
+        Guid correlationId,
         Uri serviceUrl,
-        ConsumerOptions<TMessage> options,
-        ProcessManager processManager,
-        IHandleException exceptionHandler,
-        IConnectionPool connectionPool)
+        string subscriptionName,
+        string topic,
+        IRegisterEvent eventRegister,
+        IConsumerChannel<TMessage> initialChannel,
+        IExecute executor,
+        IStateChanged<ConsumerState> state,
+        IConsumerChannelFactory<TMessage> factory)
     {
-        Topic = options.Topic;
+        _correlationId = correlationId;
         ServiceUrl = serviceUrl;
-        SubscriptionName = options.SubscriptionName;
-
-        _options = options;
-        _processManager = processManager;
-        _exceptionHandler = exceptionHandler;
-        _connectionPool = connectionPool;
-
-        _state = new StateManager<ConsumerState>(ConsumerState.Disconnected, ConsumerState.Closed, ConsumerState.ReachedEndOfTopic, ConsumerState.Faulted);
-        _cts = new CancellationTokenSource();
-        _executor = new Executor(Guid.Empty, this, _exceptionHandler);
-        _consumers = new ConcurrentDictionary<string, IConsumer<TMessage>>();
-        _messagesQueue = new ConcurrentQueue<IMessage<TMessage>>();
+        SubscriptionName = subscriptionName;
+        Topic = topic;
+        _eventRegister = eventRegister;
+        _channel = initialChannel;
+        _executor = executor;
+        _state = state;
+        _factory = factory;
+        _commandAckPool = new DefaultObjectPool<CommandAck>(new DefaultPooledObjectPolicy<CommandAck>());
         _isDisposed = 0;
 
-        _lock = new AsyncReaderWriterLock();
-
-        _ = Setup();
+        _eventRegister.Register(new ConsumerCreated(_correlationId));
     }
 
     public async ValueTask<ConsumerState> OnStateChangeTo(ConsumerState state, CancellationToken cancellationToken)
@@ -96,407 +86,140 @@ public sealed class Consumer<TMessage> : IConsumer<TMessage>, IRegisterEvent
         if (Interlocked.Exchange(ref _isDisposed, 1) != 0)
             return;
 
-        _cts.Cancel();
-        _cts.Dispose();
-
-        _state.SetState(ConsumerState.Closed);
-
-        using (_lock.ReaderLock())
-        {
-            foreach (var consumer in _consumers.Values)
-            {
-                await consumer.DisposeAsync().ConfigureAwait(false);
-            }
-        }
+        _eventRegister.Register(new ConsumerDisposed(_correlationId));
+        await DisposeChannel().ConfigureAwait(false);
     }
 
-    private async Task Setup()
+    private async ValueTask DisposeChannel()
     {
-        await Task.Yield();
-
-        try
-        {
-            await _executor.Execute(Monitor, _cts.Token).ConfigureAwait(false);
-        }
-        catch (Exception exception)
-        {
-            if (_cts.IsCancellationRequested)
-                return;
-
-            _faultException = exception;
-            _state.SetState(ConsumerState.Faulted);
-        }
-    }
-    private async Task Monitor()
-    {
-        _numberOfPartitions = await GetNumberOfPartitions(Topic, _cts.Token).ConfigureAwait(false);
-        _isPartitioned = _numberOfPartitions != 0;
-        var monitoringTasks = new List<Task<ConsumerStateChanged>>();
-
-        using (_lock.ReaderLock())
-        {
-            if (_isPartitioned)
-            {
-
-                for (var partition = 0; partition < _numberOfPartitions; ++partition)
-                {
-                    var partitionedTopicName = getPartitonedTopicName(partition);
-
-                    var consumer = CreateSubConsumer(partitionedTopicName);
-                    _ = _consumers.TryAdd(partitionedTopicName, consumer);
-                    monitoringTasks.Add(consumer.StateChangedFrom(ConsumerState.Disconnected, _cts.Token).AsTask());
-                }
-            }
-
-            else
-            {
-                var consumer = CreateSubConsumer(Topic);
-                _ = _consumers.TryAdd(Topic, consumer);
-                monitoringTasks.Add(consumer.StateChangedFrom(ConsumerState.Disconnected, _cts.Token).AsTask());
-            }
-
-            Interlocked.Exchange(ref _consumerCount, monitoringTasks.Count);
-        }
-        var activeConsumers = 0;
-        while (true)
-        {
-            await Task.WhenAny(monitoringTasks).ConfigureAwait(false);
-
-            for (var i = 0; i < monitoringTasks.Count; ++i)
-            {
-                var task = monitoringTasks[i];
-                if (!task.IsCompleted)
-                    continue;
-                var state = task.Result.ConsumerState;
-                switch (state)
-                {
-                    case ConsumerState.Active:
-                        ++activeConsumers;
-                        break;
-                    case ConsumerState.Disconnected:
-                        --activeConsumers;
-                        break;
-                    case ConsumerState.ReachedEndOfTopic:
-                        _state.SetState(ConsumerState.ReachedEndOfTopic);
-                        return;
-                    case ConsumerState.Faulted:
-                        _state.SetState(ConsumerState.Faulted);
-                        return;
-                    case ConsumerState.Unsubscribed:
-                        _state.SetState(ConsumerState.Unsubscribed);
-                        return;
-                }
-
-                monitoringTasks[i] = task.Result.Consumer.StateChangedFrom(state, _cts.Token).AsTask();
-            }
-
-            if (activeConsumers == 0)
-                _state.SetState(ConsumerState.Disconnected);
-            else if (activeConsumers == monitoringTasks.Count)
-                _state.SetState(ConsumerState.Active);
-            else
-                _state.SetState(ConsumerState.PartiallyActive);
-        }
+        await _channel.ClosedByClient(CancellationToken.None).ConfigureAwait(false);
+        await _channel.DisposeAsync().ConfigureAwait(false);
     }
 
-    private SubConsumer<TMessage> CreateSubConsumer(string topic)
-    {
-        var correlationId = Guid.NewGuid();
-        var consumerName = _options.ConsumerName ?? $"Consumer-{correlationId:N}";
+    public async ValueTask<IMessage<TMessage>> Receive(CancellationToken cancellationToken)
+        => await _executor.Execute(() => InternalReceive(cancellationToken), cancellationToken).ConfigureAwait(false);
 
-        var subscribe = new CommandSubscribe
-        {
-            ConsumerName = consumerName,
-            InitialPosition = (CommandSubscribe.InitialPositionType) _options.InitialPosition,
-            PriorityLevel = _options.PriorityLevel,
-            ReadCompacted = _options.ReadCompacted,
-            ReplicateSubscriptionState = _options.ReplicateSubscriptionState,
-            Subscription = _options.SubscriptionName,
-            Topic = topic,
-            Type = (CommandSubscribe.SubType) _options.SubscriptionType
-        };
-
-        foreach (var property in _options.SubscriptionProperties)
-        {
-            var keyValue = new KeyValue { Key = property.Key, Value = property.Value };
-            subscribe.SubscriptionProperties.Add(keyValue);
-        }
-
-        var messagePrefetchCount = _options.MessagePrefetchCount;
-        var messageFactory = new MessageFactory<TMessage>(_options.Schema);
-        var batchHandler = new BatchHandler<TMessage>(true, messageFactory);
-        var decompressorFactories = CompressionFactories.DecompressorFactories();
-
-        var factory = new ConsumerChannelFactory<TMessage>(correlationId, _processManager, _connectionPool, subscribe, messagePrefetchCount, batchHandler, messageFactory, decompressorFactories);
-        var stateManager = new StateManager<ConsumerState>(ConsumerState.Disconnected, ConsumerState.Closed, ConsumerState.ReachedEndOfTopic, ConsumerState.Faulted);
-        var initialChannel = new NotReadyChannel<TMessage>();
-        var executor = new Executor(correlationId, _processManager, _exceptionHandler);
-        var consumer = new SubConsumer<TMessage>(correlationId, ServiceUrl, _options.SubscriptionName, topic, _processManager, initialChannel, executor, stateManager, factory);
-        var process = new ConsumerProcess(correlationId, stateManager, consumer, _options.SubscriptionType == SubscriptionType.Failover);
-        _processManager.Add(process);
-        process.Start();
-        return consumer;
-    }
-
-    private async Task<uint> GetNumberOfPartitions(string topic, CancellationToken cancellationToken)
-    {
-        var connection = await _connectionPool.FindConnectionForTopic(topic, cancellationToken).ConfigureAwait(false);
-        var commandPartitionedMetadata = new PulsarApi.CommandPartitionedTopicMetadata { Topic = topic };
-        var response = await connection.Send(commandPartitionedMetadata, cancellationToken).ConfigureAwait(false);
-
-        response.Expect(PulsarApi.BaseCommand.Type.PartitionedMetadataResponse);
-
-        if (response.PartitionMetadataResponse.Response == PulsarApi.CommandPartitionedTopicMetadataResponse.LookupType.Failed)
-            response.PartitionMetadataResponse.Throw();
-
-        return response.PartitionMetadataResponse.Partitions;
-    }
-
-    private void ThrowIfDisposed()
-    {
-        if (_isDisposed != 0)
-            throw new ConsumerDisposedException(GetType().FullName!);
-    }
-    public async ValueTask Acknowledge(MessageId messageId, CancellationToken cancellationToken = default)
-    {
-        ThrowIfDisposed();
-
-        var sourceTopic = Topic;
-        if (_isPartitioned)
-        {
-            sourceTopic = getPartitonedTopicName(messageId.Partition);
-        }
-        await _executor.Execute(() =>
-                {
-                    ThrowIfNotActive();
-
-                    using (_lock.ReaderLock())
-                    {
-                        return _consumers[sourceTopic].Acknowledge(messageId, cancellationToken);
-                    }
-                }, cancellationToken)
-                .ConfigureAwait(false);
-    }
+    public async ValueTask Acknowledge(MessageId messageId, CancellationToken cancellationToken)
+        => await InternalAcknowledge(messageId, CommandAck.AckType.Individual, cancellationToken).ConfigureAwait(false);
 
     public async ValueTask AcknowledgeCumulative(MessageId messageId, CancellationToken cancellationToken)
+        => await InternalAcknowledge(messageId, CommandAck.AckType.Cumulative, cancellationToken).ConfigureAwait(false);
+
+    public async ValueTask RedeliverUnacknowledgedMessages(IEnumerable<MessageId> messageIds, CancellationToken cancellationToken)
     {
-        ThrowIfDisposed();
+        var command = new CommandRedeliverUnacknowledgedMessages();
+        command.MessageIds.AddRange(messageIds.Select(messageId => messageId.ToMessageIdData()));
+        await _executor.Execute(() => InternalRedeliverUnacknowledgedMessages(command, cancellationToken), cancellationToken).ConfigureAwait(false);
+    }
 
-        await _executor.Execute(() =>
-            {
-                ThrowIfNotActive();
+    public async ValueTask RedeliverUnacknowledgedMessages(CancellationToken cancellationToken)
+        => await RedeliverUnacknowledgedMessages(Enumerable.Empty<MessageId>(), cancellationToken).ConfigureAwait(false);
 
-                using (_lock.ReaderLock())
-                {
-                    var sourceTopic = Topic;
-                    if (_isPartitioned)
-                    {
-                        sourceTopic = getPartitonedTopicName(messageId.Partition);
-                    }
-                    return _consumers[sourceTopic].AcknowledgeCumulative(messageId, cancellationToken);
-                }
-            }, cancellationToken)
-            .ConfigureAwait(false);
+    public async ValueTask Unsubscribe(CancellationToken cancellationToken)
+    {
+        var unsubscribe = new CommandUnsubscribe();
+        await _executor.Execute(() => InternalUnsubscribe(unsubscribe, cancellationToken), cancellationToken).ConfigureAwait(false);
+    }
+
+    public async ValueTask Seek(MessageId messageId, CancellationToken cancellationToken)
+    {
+        var seek = new CommandSeek { MessageId = messageId.ToMessageIdData() };
+        await _executor.Execute(() => InternalSeek(seek, cancellationToken), cancellationToken).ConfigureAwait(false);
+    }
+
+    public async ValueTask Seek(ulong publishTime, CancellationToken cancellationToken)
+    {
+        var seek = new CommandSeek { MessagePublishTime = publishTime };
+        await _executor.Execute(() => InternalSeek(seek, cancellationToken), cancellationToken).ConfigureAwait(false);
     }
 
     public async ValueTask<MessageId> GetLastMessageId(CancellationToken cancellationToken)
     {
-        ThrowIfDisposed();
-
         var getLastMessageId = new CommandGetLastMessageId();
-        return await _executor.Execute(() => GetLastMessageId(getLastMessageId, cancellationToken), cancellationToken).ConfigureAwait(false);
+        return await _executor.Execute(() => InternalGetLastMessageId(getLastMessageId, cancellationToken), cancellationToken).ConfigureAwait(false);
     }
 
-    private async ValueTask<MessageId> GetLastMessageId(CommandGetLastMessageId command, CancellationToken cancellationToken)
+    private void Guard()
     {
-        ThrowIfNotActive();
+        if (_isDisposed != 0)
+            throw new ConsumerDisposedException(GetType().FullName!);
 
-        if (_isPartitioned)
+        if (_faultException is not null)
+            throw new ConsumerFaultedException(_faultException);
+    }
+
+    public async Task EstablishNewChannel(CancellationToken cancellationToken)
+    {
+        var channel = await _executor.Execute(() => _factory.Create(cancellationToken), cancellationToken).ConfigureAwait(false);
+
+        var oldChannel = _channel;
+        if (oldChannel is not null)
+            await oldChannel.DisposeAsync().ConfigureAwait(false);
+
+        _channel = channel;
+    }
+
+    public async ValueTask CloseChannel(CancellationToken cancellationToken)
+        => await _channel.ClosedByClient(cancellationToken).ConfigureAwait(false);
+
+    public async ValueTask ChannelFaulted(Exception exception)
+    {
+        _faultException = exception;
+        await DisposeChannel().ConfigureAwait(false);
+    }
+
+    private async ValueTask InternalAcknowledge(CommandAck command, CancellationToken cancellationToken)
+    {
+        Guard();
+        await _channel.Send(command, cancellationToken).ConfigureAwait(false);
+    }
+
+    private async ValueTask InternalRedeliverUnacknowledgedMessages(CommandRedeliverUnacknowledgedMessages command, CancellationToken cancellationToken)
+    {
+        Guard();
+        await _channel.Send(command, cancellationToken).ConfigureAwait(false);
+    }
+
+    private async ValueTask<MessageId> InternalGetLastMessageId(CommandGetLastMessageId command, CancellationToken cancellationToken)
+    {
+        Guard();
+        return await _channel.Send(command, cancellationToken).ConfigureAwait(false);
+    }
+
+    private async Task InternalSeek(CommandSeek command, CancellationToken cancellationToken)
+    {
+        Guard();
+        await _channel.Send(command, cancellationToken).ConfigureAwait(false);
+    }
+
+    private async ValueTask<IMessage<TMessage>> InternalReceive(CancellationToken cancellationToken)
+    {
+        Guard();
+        return await _channel.Receive(cancellationToken).ConfigureAwait(false);
+    }
+
+    private async ValueTask InternalUnsubscribe(CommandUnsubscribe command, CancellationToken cancellationToken)
+    {
+        Guard();
+        await _channel.Send(command, cancellationToken).ConfigureAwait(false);
+    }
+
+    private async ValueTask InternalAcknowledge(MessageId messageId, CommandAck.AckType ackType, CancellationToken cancellationToken)
+    {
+        var commandAck = _commandAckPool.Get();
+        commandAck.Type = ackType;
+        if (commandAck.MessageIds.Count == 0)
+            commandAck.MessageIds.Add(messageId.ToMessageIdData());
+        else
+            commandAck.MessageIds[0].MapFrom(messageId);
+
+        try
         {
-            throw new NotImplementedException("GetLastMessageId is not implemented for partitioned topics");
+            await _executor.Execute(() => InternalAcknowledge(commandAck, cancellationToken), cancellationToken).ConfigureAwait(false);
         }
-        using (_lock.ReaderLock())
+        finally
         {
-            return await _consumers.First().Value.GetLastMessageId(cancellationToken).ConfigureAwait(false);
+            _commandAckPool.Return(commandAck);
         }
     }
-
-    public async ValueTask<IMessage<TMessage>> Receive(CancellationToken cancellationToken = default)
-    {
-        ThrowIfDisposed();
-
-        return await _executor.Execute(() => ReceiveMessage(cancellationToken), cancellationToken).ConfigureAwait(false);
-    }
-
-    private async ValueTask<IMessage<TMessage>> ReceiveMessage(CancellationToken cancellationToken)
-    {
-        ThrowIfNotActive();
-
-        using (_lock.ReaderLock())
-        {
-            if (_messagesQueue.TryDequeue(out var message))
-            {
-                return message;
-            }
-            var cts = new CancellationTokenSource();
-            var linkedCts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken, cts.Token);
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                var done = false;
-
-                Task<IMessage<TMessage>>[] receiveTasks = _consumers.Values.Select(consumer => consumer.Receive(linkedCts.Token).AsTask()).ToArray();
-                await Task.WhenAny(receiveTasks).ConfigureAwait(false);
-
-                try
-                {
-                    receiveTasks.Where(t => t.IsCompleted).ToList().ForEach(t =>
-                        {
-                            if (t.Result == null)
-                            {
-                                return;
-                            }
-
-                            done = true;
-                            _messagesQueue.Enqueue(t.Result);
-                        }
-                    );
-                }
-                catch (Exception exception)
-                {
-                    if (linkedCts.IsCancellationRequested)
-                    {
-                        cts.Cancel();
-                    }
-                    else
-                    {
-                        throw exception;
-                    }
-                }
-                if (done)
-                {
-                    break;
-                }
-            }
-            cts.Cancel();
-            cts.Dispose();
-            _messagesQueue.TryDequeue(out var result);
-            return result;
-        }
-    }
-
-    public async ValueTask RedeliverUnacknowledgedMessages(IEnumerable<MessageId> messageIds, CancellationToken cancellationToken = default)
-    {
-        ThrowIfDisposed();
-
-        await _executor.Execute<ValueTask>(async () =>
-        {
-            ThrowIfNotActive();
-
-            using (_lock.ReaderLock())
-            {
-                var tasks = messageIds.Select(n =>
-                {
-                    var sourceTopic = Topic;
-                    if (_isPartitioned)
-                    {
-                        sourceTopic = getPartitonedTopicName(n.Partition);
-                    }
-                    return _consumers[getPartitonedTopicName(n.Partition)].RedeliverUnacknowledgedMessages(cancellationToken).AsTask();
-                });
-                await Task.WhenAll(tasks).ConfigureAwait(false);
-            }
-        }, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async ValueTask RedeliverUnacknowledgedMessages(CancellationToken cancellationToken = default)
-    {
-        ThrowIfDisposed();
-
-        await _executor.Execute<ValueTask>(async () =>
-        {
-            ThrowIfNotActive();
-
-            using (_lock.ReaderLock())
-            {
-                var tasks = _consumers.Values.Select(consumer =>
-                    consumer.RedeliverUnacknowledgedMessages(cancellationToken).AsTask()
-                );
-                await Task.WhenAll(tasks).ConfigureAwait(false);
-            }
-        }, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async ValueTask Seek(MessageId messageId, CancellationToken cancellationToken = default)
-    {
-        ThrowIfDisposed();
-
-        await _executor.Execute<ValueTask>(async () =>
-        {
-            ThrowIfNotActive();
-
-            using (_lock.ReaderLock())
-            {
-                if (messageId.Equals(null))
-                {
-                    throw new ArgumentException("Illegal messageId cannot be null");
-                }
-
-                var tasks = _consumers.Values.Select(consumer =>
-                    consumer.Seek(messageId, cancellationToken).AsTask()
-                );
-                await Task.WhenAll(tasks.ToArray()).ConfigureAwait(false);
-                _messagesQueue = new ConcurrentQueue<IMessage<TMessage>>();
-            }
-        }, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async ValueTask Seek(ulong publishTime, CancellationToken cancellationToken = default)
-    {
-        ThrowIfDisposed();
-
-        await _executor.Execute<ValueTask>(async () =>
-        {
-            ThrowIfNotActive();
-
-            using (_lock.ReaderLock())
-            {
-                var tasks = _consumers.Values.Select(consumer =>
-                    consumer.Seek(publishTime, cancellationToken).AsTask()
-                );
-                await Task.WhenAll(tasks.ToArray()).ConfigureAwait(false);
-                _messagesQueue = new ConcurrentQueue<IMessage<TMessage>>();
-            }
-        }, cancellationToken).ConfigureAwait(false);
-    }
-
-    public async ValueTask Unsubscribe(CancellationToken cancellationToken)
-    {
-        ThrowIfDisposed();
-
-        await _executor.Execute<ValueTask>(async () =>
-        {
-            ThrowIfNotActive();
-
-            using (_lock.ReaderLock())
-            {
-                var tasks = _consumers.Values.Select(consumer =>
-                    consumer.Unsubscribe(cancellationToken).AsTask()
-                );
-                await Task.WhenAll(tasks).ConfigureAwait(false);
-            }
-        }, cancellationToken).ConfigureAwait(false);
-    }
-
-    private string getPartitonedTopicName(int partitionNumber)
-    {
-        return $"{Topic}-partition-{partitionNumber}";
-    }
-
-    private void ThrowIfNotActive()
-    {
-        if (_state.CurrentState != ConsumerState.Active)
-            throw new ConsumerNotActiveException("The consumer is not yet activated.");
-    }
-
-    public void Register(IEvent @event) { }
 }
